+++ tests/client_configuration_tests.rs
@@ -1,8 +1,5 @@
-use pass_it_on::notifications::Key;
 use pass_it_on::ClientConfiguration;
 use pass_it_on::Error;
-
-const VALID_KEY: &[u8; 32] = b"sdfsf4633ghf44dfhdfhQdhdfhewaasg";
 
 #[test]
 #[cfg(unix)]
@@ -54,25 +51,8 @@ fn client_valid_config_windows() {
 }
 
 #[test]
-fn server_invalid_key_length() {
-    let config = ClientConfiguration::try_from(
-        r#"
-    [client]
-    key = "asdffdsa12346785"
-
-    [[client.interface]]
-    type = "http"
-    port = 8080
-
-"#,
-    );
-
-    assert_eq!(config.unwrap_err().to_string(), Error::InvalidKeyLength(16).to_string())
-}
-
-#[test]
 fn interface_not_defined() {
-    let config = ClientConfiguration::new(Key::from_bytes(VALID_KEY), Vec::new());
+    let config = ClientConfiguration::new("test key", Vec::new());
 
     assert_eq!(config.unwrap_err().to_string(), Error::MissingInterface.to_string())
 }